Allow client to run multiple times
Refactor the Client code so that it can support multiple runs.  The
`run` command now takes an message argument which was originally passed
the `Client::new`.
